Refactor: type 폴더 -> enumerate로 변경
