+++ payment-api/src/main/java/org/ecommerce/paymentapi/entity/enumerate/PaymentStatus.java
@@ -1,17 +1,13 @@
-package org.ecommerce.paymentapi.entity.type;
+package org.ecommerce.paymentapi.entity.enumerate;
 
 import org.ecommerce.common.utils.mapper.EnumMapperType;
 
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
-public enum RefundStatus implements EnumMapperType {
-	PENDING("환불 처리전"),
-	PROCESSING("환불 진행중"),
-	COMPLETED("환불 처리완료"),
-	FAILED("환불 실패"),
-	CANCLLED("환불 취소");
+public enum PaymentStatus implements EnumMapperType {
+	PAYMENT("결제"),
+	REFUND("환불");
 
 	private final String title;
 
@@ -19,6 +15,7 @@ public enum RefundStatus implements EnumMapperType {
 	public String getCode() {
 		return name();
 	}
+
 	@Override
 	public String getTitle() {
 		return this.title;
