+++ payment-api/src/main/java/org/ecommerce/paymentapi/entity/enumerate/ProcessStatus.java
@@ -1,15 +1,16 @@
-package org.ecommerce.paymentapi.entity.type;
+package org.ecommerce.paymentapi.entity.enumerate;
 
 import org.ecommerce.common.utils.mapper.EnumMapperType;
 
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
-public enum PaymentStatus implements EnumMapperType {
-	PAYMENT("결제"),
-	REFUND("환불");
-
+public enum ProcessStatus implements EnumMapperType {
+	PENDING("처리 전"),
+	IN_PROGRESS("진행 중"),
+	FAILED("실패"),
+	CANCELLED("취소"),
+	COMPLETED("완료");
 	private final String title;
 
 	@Override
@@ -19,6 +20,6 @@ public enum PaymentStatus implements EnumMapperType {
 
 	@Override
 	public String getTitle() {
-		return this.title;
+		return title;
 	}
 }
