+++ payment-api/src/main/java/org/ecommerce/paymentapi/entity/enumerate/RefundStatus.java
@@ -1,14 +1,16 @@
-package org.ecommerce.paymentapi.entity.type;
+package org.ecommerce.paymentapi.entity.enumerate;
 
 import org.ecommerce.common.utils.mapper.EnumMapperType;
 
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
-public enum BeanPayStatus implements EnumMapperType {
-	DEPOSIT("충전"),
-	WITHDRAW("출금");
+public enum RefundStatus implements EnumMapperType {
+	PENDING("환불 처리전"),
+	PROCESSING("환불 진행중"),
+	COMPLETED("환불 처리완료"),
+	FAILED("환불 실패"),
+	CANCLLED("환불 취소");
 
 	private final String title;
 
@@ -16,7 +18,6 @@ public enum BeanPayStatus implements EnumMapperType {
 	public String getCode() {
 		return name();
 	}
-
 	@Override
 	public String getTitle() {
 		return this.title;
