+++ payment-api/src/test/java/org/ecommerce/paymentapi/repository/BeanPayRepositoryTest.java
@@ -1,24 +1,29 @@
 package org.ecommerce.paymentapi.repository;
 
-import static org.ecommerce.paymentapi.entity.type.Role.*;
+import static org.ecommerce.paymentapi.entity.enumerate.Role.*;
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.time.LocalDateTime;
 import java.util.Optional;
 
+import org.ecommerce.paymentapi.config.QueryDslConfig;
 import org.ecommerce.paymentapi.entity.BeanPay;
-import org.ecommerce.paymentapi.entity.type.Role;
+import org.ecommerce.paymentapi.entity.enumerate.Role;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 
+
+@Import(QueryDslConfig.class)
 @ActiveProfiles("test")
 @DataJpaTest
 public class BeanPayRepositoryTest {
 	@Autowired
 	private BeanPayRepository beanPayRepository;
+
 	@Nested
 	class findBeanPayByUserIdAndRole {
 		@Test
